requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
PyMySQL>=1.1.0
cryptography>=41.0.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
openai>=1.0.0
